@@ -4,7 +4,7 @@ version: 2.0.6
 homepage: https://github.com/namanh11611/curved_labeled_navigation_bar
 
 environment:
-  sdk: ">=2.12.0 <4.0.0"
+  sdk: ">=2.17.0 <4.0.0"
   flutter: ">=3.10.0"
 
 dependencies:
