@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class CurvedNavigationBarItem {
   /// Icon of [CurvedNavigationBarItem].
-  final Widget child;
+  final IconData child;
 
   /// Text of [CurvedNavigationBarItem].
   final String? label;
